+++ lib/skeem/grammar.rb
@@ -11,14 +11,14 @@ module Skeem
     # Delimiters, separators...
     # add_terminals('APOSTROPHE', 'BACKQUOTE')
     add_terminals('LPAREN', 'RPAREN')
-    # add_terminals('PERIOD')
+    add_terminals('PERIOD')
 
     # Literal values...
     add_terminals('BOOLEAN', 'INTEGER', 'REAL')
     add_terminals('STRING_LIT', 'IDENTIFIER')
 
     # Keywords...
-    add_terminals('DEFINE', 'IF')
+    add_terminals('DEFINE', 'IF', 'LAMBDA')
 
     rule('program' => 'cmd_or_def_plus').as 'main'
     rule('cmd_or_def_plus' => 'cmd_or_def_plus cmd_or_def').as 'multiple_cmd_def'
@@ -30,6 +30,7 @@ module Skeem
     rule('expression' =>  'IDENTIFIER').as 'variable_reference'
     rule 'expression' =>  'literal'
     rule 'expression' =>  'procedure_call'
+    rule 'expression' =>  'lambda_expression'
     rule 'expression' =>  'conditional'
     rule 'literal' => 'self-evaluating'
     rule 'self-evaluating' => 'BOOLEAN'
@@ -41,6 +42,20 @@ module Skeem
     rule('operand_plus' => 'operand').as 'last_operand'
     rule 'operator' => 'expression'
     rule 'operand' => 'expression'
+    rule('lambda_expression' => 'LPAREN LAMBDA formals body RPAREN').as 'lambda_expression'
+    rule('formals' => 'LPAREN identifier_star RPAREN').as 'identifiers_as_formals'
+    rule 'formals' => 'IDENTIFIER'
+    rule 'formals' => 'LPAREN identifier_plus PERIOD IDENTIFIER RPAREN'
+    rule('identifier_star' => 'identifier_star IDENTIFIER').as 'identifier_star'
+    rule('identifier_star' => []).as 'no_identifier_yet'
+    rule 'identifier_plus' => 'identifier_plus IDENTIFIER'
+    rule 'identifier_plus' => 'IDENTIFIER'
+    rule('body' => 'definition_star sequence').as 'body'
+    rule 'definition_star' => 'definition_star definition'
+    rule 'definition_star' => []
+    rule('sequence' => 'command_star expression').as 'sequence'
+    rule('command_star' => 'command_star command').as 'multiple_commands'
+    rule('command_star' => []).as 'no_command_yet'
     rule('conditional' => 'LPAREN IF test consequent alternate RPAREN').as 'conditional'
     rule 'test' => 'expression'
     rule 'consequent' => 'expression'