+++ lib/skeem/primitive_procedure.rb
@@ -5,11 +5,12 @@ module Skeem
     attr_reader(:identifier)
     attr_reader(:code)
 
-    def initialize(anId, aLambda)
+    def initialize(anId, aRubyLambda)
       @identifier = anId.kind_of?(String) ? SkmIdentifier.create(anId) : anId
-      @code = aLambda
+      @code = aRubyLambda
     end
 
+    # Arguments are positional in a primitive procedure.
     def call(aRuntime, aProcedureCall)
       args = aProcedureCall.operands
       return @code.call(aRuntime, args)