+++ lib/skeem/s_expr_nodes.rb
@@ -36,24 +36,38 @@ module Skeem
     def integer?
       false
     end
-    
+
     def boolean?
       false
     end
 
     def string?
       false
-    end    
+    end
 
     def symbol?
       false
-    end    
+    end
 
     # Abstract method.
     # Part of the 'visitee' role in Visitor design pattern.
     # @param _visitor[ParseTreeVisitor] the visitor
     def accept(_visitor)
       raise NotImplementedError
+    end
+
+   def inspect()
+      raise NotImplementedError
+    end
+
+    protected
+
+    def inspect_prefix
+      "<#{self.class.name}: "
+    end
+
+    def inspect_suffix
+      '>'
     end
   end # struct
 
@@ -75,17 +89,16 @@ module Skeem
       return lightweight
     end
 
-    # This method can be overriden
-    def init_value(aValue)
-      @value = aValue
-    end
-
     def symbol()
       token.terminal
     end
 
     def evaluate(_runtime)
       return self
+    end
+
+    def inspect()
+      inspect_prefix + value.to_s + inspect_suffix
     end
 
     def done!()
@@ -97,12 +110,17 @@ module Skeem
     def accept(aVisitor)
       aVisitor.visit_terminal(self)
     end
+
+    # This method can be overriden
+    def init_value(aValue)
+      @value = aValue
+    end
   end # class
 
   class SkmBoolean < SkmTerminal
     def boolean?
       true
-    end  
+    end
   end # class
 
   class SkmNumber < SkmTerminal
@@ -128,10 +146,10 @@ module Skeem
     def init_value(aValue)
       super(aValue.dup)
     end
-    
+
     def string?
       true
-    end     
+    end
   end # class
 
   class SkmIdentifier < SkmTerminal
@@ -139,10 +157,10 @@ module Skeem
     def init_value(aValue)
       super(aValue.dup)
     end
-    
+
     def symbol?
       true
-    end     
+    end
   end # class
 
   class SkmReserved < SkmIdentifier
@@ -153,7 +171,7 @@ module Skeem
     attr_accessor(:members)
     extend Forwardable
 
-    def_delegators :@members, :first, :length, :empty?
+    def_delegators :@members, :each, :first, :length, :empty?
 
     def initialize(theMembers)
       super(nil)
@@ -202,8 +220,17 @@ module Skeem
       # Do nothing
     end
 
+    def inspect()
+      result = inspect_prefix
+      members.each { |elem| result << elem.inspect + ', ' }
+      result.sub!(/, $/, '')
+      result << inspect_suffix
+      result
+    end
+
     alias children members
     alias subnodes members
+    alias rest tail
   end # class
 
   class SkmDefinition < SkmElement
@@ -219,13 +246,32 @@ module Skeem
     def evaluate(aRuntime)
       var_key = variable.evaluate(aRuntime)
       aRuntime.define(var_key, self)
+      expression.evaluate(aRuntime) if expression.kind_of?(SkmLambda)
       self
     end
+
+    # call method should only invoked when the expression is a SkmLambda
+    def call(aRuntime, aProcedureCall)
+      unless expression.kind_of?(SkmLambda)
+        err_msg = "Expected a SkmLambda instead of #{expression.class}"
+        raise StandardError, err_msg
+      end
+      expression.call(aRuntime, aProcedureCall)
+    end
+
+    def inspect
+      result = inspect_prefix
+      result << variable.inspect
+      result << ', '
+      result << expression.inspect
+      result << inspect_suffix
+      result
+    end
   end # class
-  
+
   class SkmVariableReference  < SkmElement
     attr_reader :variable
-    
+
     def initialize(aPosition, aVariable)
       super(aPosition)
       @variable = aVariable
@@ -236,10 +282,10 @@ module Skeem
       unless aRuntime.include?(var_key.value)
         err = StandardError
         key = var_key.kind_of?(SkmIdentifier) ? var_key.value : var_key
-        err_msg = "Unknown variable '#{key}'"
+        err_msg = "Unbound variable: '#{key}'"
         raise err, err_msg
       end
-      definition = aRuntime.environment.bindings[var_key.value]
+      definition = aRuntime.environment.fetch(var_key.value)
       result = definition.expression.evaluate(aRuntime)
     end
 
@@ -248,7 +294,12 @@ module Skeem
     def value()
       variable.value
     end
-  end
+
+    def inspect
+      result = inspect_prefix + variable.inspect + inspect_suffix
+      result
+    end
+  end # class
 
   class ProcedureCall < SkmElement
     attr_reader :operator
@@ -273,13 +324,19 @@ module Skeem
         err_msg = "Unknown procedure '#{key}'"
         raise err, err_msg
       end
-      procedure = aRuntime.environment.bindings[var_key.value]
+      procedure = aRuntime.environment.fetch(var_key.value)
       result = procedure.call(aRuntime, self)
+    end
+
+    def inspect
+      result = inspect_prefix + operator.inspect + ', '
+      result << operands.inspect + inspect_suffix
+      result
     end
 
     alias children operands
   end # class
-  
+
   class SkmCondition < SkmElement
     attr_reader :test
     attr_reader :consequent
@@ -301,6 +358,95 @@ module Skeem
       else
         condition_result = consequent.evaluate(aRuntime)
       end
+    end
+
+    def inspect
+      result = inspect_prefix + '@test ' + test.inspect + ', '
+      result << '@consequent ' + consequent.inspect + ', '
+      result << '@alternate ' + alternate.inspect + inspect_suffix
+      result
+    end
+  end # class
+  
+=begin
+<Skeem::SkmLambda: 
+  @formals [<Skeem::SkmIdentifier: x>]
+  @definitions nil
+  @sequence <Skeem::SkmList: 
+    <Skeem::SkmCondition: 
+      @test <Skeem::ProcedureCall: <Skeem::SkmIdentifier: <>, 
+        <Skeem::SkmList: <Skeem::SkmVariableReference: <Skeem::SkmIdentifier: x>>, 
+        <Skeem::SkmInteger: 0>>>, 
+      @consequent <Skeem::ProcedureCall: 
+        <Skeem::SkmIdentifier: ->, 
+        <Skeem::SkmList: <Skeem::SkmVariableReference: <Skeem::SkmIdentifier: x>>>>,
+      @alternate <Skeem::SkmVariableReference: <Skeem::SkmIdentifier: x>>>>>
+=end
+
+  class SkmLambda < SkmElement
+    # @!attribute [r] formals
+    # @return [Array<SkmIdentifier>] the argument names
+    attr_reader :formals
+    attr_reader :definitions
+    attr_reader :sequence
+
+    def initialize(aPosition, theFormals, aBody)
+      super(aPosition)
+      @formals = theFormals
+      @definitions = aBody[:defs]
+      @sequence = aBody[:sequence]
+    end
+
+    def evaluate(aRuntime)
+      formals.map! { |param| param.evaluate(aRuntime) }
+    end
+
+    def call(aRuntime, aProcedureCall)
+      aRuntime.nest
+      bind_locals(aRuntime, aProcedureCall)
+      # TODO remove next line
+      # puts aRuntime.environment.inspect
+      result = evaluate_defs(aRuntime)
+      result = evaluate_sequence(aRuntime)
+      aRuntime.unnest
+
+      result
+    end
+
+    def inspect
+      result = inspect_prefix + '@formals ' + formals.inspect + ', '
+      result << '@definitions ' + definitions.inspect + ', '
+      result << '@sequence ' + sequence.inspect + inspect_suffix
+      result
+    end
+
+    private
+
+    def bind_locals(aRuntime, aProcedureCall)
+      arguments =  aProcedureCall.operands.members
+      formals.each_with_index do |arg_name, index|
+        arg = arguments[index]
+        if arg.nil?
+          p aProcedureCall.operands.members
+          raise StandardError, "Unbound variable: '#{arg_name.value}'"
+        end
+
+        a_def = SkmDefinition.new(position, arg_name, arg)
+        a_def.evaluate(aRuntime)
+      end
+    end
+
+    def evaluate_defs(aRuntime)
+      definitions.each { |a_def| a_def.evaluate(runtime) }
+    end
+
+    def evaluate_sequence(aRuntime)
+      result = nil
+      if sequence
+        sequence.each { |cmd| result = cmd.evaluate(aRuntime) }
+      end
+
+      result
     end
   end # class
 end # module