+++ spec/skeem/environment_spec.rb
@@ -4,12 +4,12 @@ require_relative '../../lib/skeem/environment' # Load the class under test
 module Skeem
   describe Environment do
     context 'Initialization:' do
-      it 'should be initialized without argument' do
+      it 'should be initialized with opional argument' do
         expect { Environment.new() }.not_to raise_error
       end
 
-      it 'should have no bindings' do
-        expect(subject.bindings).to be_empty
+      it 'should have no default bindings' do
+        expect(subject).to be_empty
       end
     end # context
 
@@ -17,11 +17,53 @@ module Skeem
       it 'should add entries' do
         entry = double('dummy')
         subject.define('dummy', entry)
-        expect(subject.bindings.size).to eq(1) 
+        expect(subject.size).to eq(1)
         expect(subject.bindings['dummy']).not_to be_nil
         expect(subject.bindings['dummy']).to eq(entry)
       end
+
+      it 'should know whether it is empty' do
+        # Case 1: no entry
+        expect(subject.empty?).to be_truthy
+
+        # Case 2: existing entry
+        entry = double('dummy')
+        subject.define('dummy', entry)
+        expect(subject.empty?).to be_falsey
+
+        # Case 3: entry defined in outer environment
+        nested = Environment.new(subject)
+        expect(nested.empty?).to be_falsey
+      end
+
+      it 'should retrieve entries' do
+        # Case 1: non-existing entry
+        expect(subject.fetch('dummy')).to be_nil
+
+        # Case 2: existing entry
+        entry = double('dummy')
+        subject.define('dummy', entry)
+        expect(subject.fetch('dummy')).to eq(entry)
+
+        # Case 3: entry defined in outer environment
+        nested = Environment.new(subject)
+        expect(nested.fetch('dummy')).to eq(entry)
+      end
+
+      it 'should know the total number of bindings' do
+        # Case 1: non-existing entry
+        expect(subject.size).to be_zero
+
+        # Case 2: existing entry
+        entry = double('dummy')
+        subject.define('dummy', entry)
+        expect(subject.size).to eq(1)
+
+        # Case 3: entry defined in outer environment
+        nested = Environment.new(subject)
+        expect(nested.size).to eq(1)
+      end
     end # context
-    
+
   end # describe
 end # module